requirements.txt
slack_sdk
python-dotenv
fastapi
uvicorn
pydantic-settings
psycopg2-binary
openai